+++ daemon/internal/subscription/fetch_test.go
@@ -2,15 +2,8 @@ package subscription
 
 import (
 	"context"
-	"crypto/rand"
-	"crypto/rsa"
-	"crypto/x509"
-	"encoding/pem"
 	"errors"
-	"math/big"
 	"net"
-	"os"
-	"path/filepath"
 	"strings"
 	"testing"
 	"time"
@@ -143,37 +136,6 @@ func TestBootstrapResolverIgnoresSystemLoopbackDNSServer(t *testing.T) {
 	}
 	_ = conn.Close()
 	<-done
-}
-
-func TestRootCAPoolFromDirsLoadsAndroidPEMCert(t *testing.T) {
-	key, err := rsa.GenerateKey(rand.Reader, 2048)
-	if err != nil {
-		t.Fatal(err)
-	}
-	template := &x509.Certificate{
-		SerialNumber: big.NewInt(1),
-		NotBefore:    time.Now().Add(-time.Hour),
-		NotAfter:     time.Now().Add(time.Hour),
-		IsCA:         true,
-	}
-	der, err := x509.CreateCertificate(rand.Reader, template, template, &key.PublicKey, key)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	dir := t.TempDir()
-	certPEM := pem.EncodeToMemory(&pem.Block{
-		Type:  "CERTIFICATE",
-		Bytes: der,
-	})
-	if err := os.WriteFile(filepath.Join(dir, "test.0"), certPEM, 0o644); err != nil {
-		t.Fatal(err)
-	}
-
-	pool := rootCAPoolFromDirs(x509.NewCertPool(), []string{dir})
-	if len(pool.Subjects()) == 0 {
-		t.Fatal("expected Android-style PEM cert file to be loaded")
-	}
 }
 
 func TestClassifyError(t *testing.T) {